+++ src/visuals-api.d.ts
@@ -16,13 +16,20 @@ declare namespace powerbi {
         Append = 1,
         Segment = 2,
     }
+    /** 
+     * Binary system is used for types flags managing, so VisualUpdateType value
+     * represents a combination of all received types
+     */
     const enum VisualUpdateType {
-        Data = 2,
-        Resize = 4,
-        ViewMode = 8,
-        Style = 16,
-        ResizeEnd = 32,
-        All = 62,
+        Data = 1 << 1,
+        Resize = 1 << 2,
+        ViewMode = 1 << 3,
+        Style = 1 << 4,
+        ResizeEnd = 1 << 5,
+        FormattingSubSelectionChange = 1 << 6,
+        FormatModeChange = 1 << 7,
+        FilterOptionsChange = 1 << 8,
+        All = Data | Resize | ViewMode | Style | ResizeEnd | FormattingSubSelectionChange | FormatModeChange | FilterOptionsChange
     }
     const enum VisualPermissions {
     }