+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 5.3.0
 * SelectionId's update-fix for matrix dataView.  
 *Note: the selectionId's core data might change therefore a persisted selectionIds/identityIndex using an older API version might not be relevant in matrix dataView.*
-* `downloadService`: Adds a new method `exportVisualsContentExtended` which returns expanded reuslt information of the download.      
+* `downloadService`: Adds a new method `exportVisualsContentExtended` which returns expanded reuslt information of the download.
+* Extended `VisualUpdateType` enum with 3 new types: `FormattingSubSelectionChange`, `FormatModeChange` and `FilterOptionsChange`.
 
 ## 5.2.0
 * Adds `dataReductionCustomization` into schema for capabilities.json, this capability offers a declarative way of customizing some data reduction behavior at query generation time.