+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log - Power BI Custom Visuals API
 
+## 5.4.0
+* Adds `isDataFilterApplied` into DataViewMetadata, this capability offers a boolean value of whether any filter applied affects the visual.
+
 # 5.2.0
 * Adds `dataReductionCustomization` into schema for capabilities.json, this capability offers a declarative way of customizing some data reduction behavior at query generation time.